+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Fernschach-Verwaltung Changelog
 
+## Version 0.3.4 (2022-05-13)
+
+* Add: tl_fernschach_spieler zusätzliche Felder anhand Stammdaten-Datenbank angelegt -> anrede, klassenberechtigung, telefax1, telefax2, streichung, briefanrede, gastNummer, servertesterNummer, fremdspielerNummer, zuzug, adresszusatz, verein, status
+* Change: ImportSpieler mit neuen Feldern versehen und Überschreibfunktionen angepaßt
+
 ## Version 0.3.3 (2022-05-13)
 
 * Change: Bewerbungen in Turnier-Einstellungen nur anzeigen, wenn Bewerbungen erlaubt sind